Reorganize package structure.
TNR-15
