+++ setup.py
@@ -1,0 +1,11 @@
+from distutils.core import setup
+
+setup(
+    name='VacasaConnect',
+    version='0.1.0',
+    description='A Python 3.6 SDK for the connect.vacasa.com API.',
+    packages=['vacasa'],
+    url='https://github.com/Vacasa/python-vacasa-connect-sdk',
+    license='MIT',
+    long_description=open('README.md').read()
+)
